Moving test_base.py to the cannonical test dir.
